backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==23.0.0
google-genai==1.11.0
google-generativeai==0.8.3
